Add helper functions in gpu4pyscf.gto.mole
