+++ gpu4pyscf/scf/jk.py
@@ -24,14 +24,14 @@ import math
 import numpy as np
 import cupy as cp
 import scipy.linalg
-from pyscf.gto import (ANG_OF, ATOM_OF, NPRIM_OF, NCTR_OF, PTR_COORD, PTR_COEFF,
-                       PTR_EXP, gto_norm)
+from pyscf.gto import ANG_OF, ATOM_OF, NPRIM_OF, NCTR_OF, PTR_COORD, PTR_COEFF
 from pyscf import lib
 from pyscf.scf import _vhf
 from pyscf import __config__
 from gpu4pyscf.lib.cupy_helper import load_library, condense, sandwich_dot, transpose_sum
 from gpu4pyscf.__config__ import props as gpu_specs
 from gpu4pyscf.lib import logger
+from gpu4pyscf.gto.mole import group_basis
 
 __all__ = [
     'get_jk', 'get_j',
@@ -55,9 +55,6 @@ GOUT_WIDTH = 42
 SHM_SIZE = getattr(__config__, 'GPU_SHM_SIZE',
                    int(gpu_specs['sharedMemPerBlockOptin']//9)*8)
 THREADS = 256
-
-# TODO: test different size for L2 cache efficiency
-NAO_IN_GROUP = 1500
 
 def get_jk(mol, dm, hermi=0, vhfopt=None, with_j=True, with_k=True, verbose=None):
     '''Compute J, K matrices
@@ -380,7 +377,7 @@ def get_j(mol, dm, hermi=0, vhfopt=None, verbose=None):
 
 class _VHFOpt:
     def __init__(self, mol, cutoff=1e-13):
-        self.mol, self.coeff = basis_seg_contraction(mol)
+        self.orig_mol = mol
         self.direct_scf_tol = cutoff
         self.uniq_l_ctr = None
         self.l_ctr_offsets = None
@@ -390,71 +387,14 @@ class _VHFOpt:
         self.tile = TILE
 
     def build(self, group_size=None, verbose=None):
-        mol = self.mol
+        mol = self.orig_mol
         log = logger.new_logger(mol, verbose)
         cput0 = log.init_timer()
-        # Sort basis according to angular momentum and contraction patterns so
-        # as to group the basis functions to blocks in GPU kernel.
-        l_ctrs = mol._bas[:,[ANG_OF, NPRIM_OF]]
-        # Ensure the more contracted Gaussians being accessed first
-        l_ctrs_descend = l_ctrs.copy()
-        l_ctrs_descend[:,1] = -l_ctrs[:,1]
-        uniq_l_ctr, where, inv_idx, l_ctr_counts = np.unique(
-            l_ctrs_descend, return_index=True, return_inverse=True, return_counts=True, axis=0)
-        uniq_l_ctr[:,1] = -uniq_l_ctr[:,1]
-
-        nao_orig = self.coeff.shape[1]
-        ao_loc = mol.ao_loc
-        coeff = np.split(self.coeff, ao_loc[1:-1], axis=0)
-
-        l_ctr_counts_orig = l_ctr_counts.copy()
-        pad_inv_idx = []
-        pad_bas = []
-        env_ptr = mol._env.size
-        tile = self.tile
-        # for each pattern, padding basis to the end of mol._bas, ensure alignment to TILE
-        for n, (l_ctr, m, counts) in enumerate(zip(uniq_l_ctr, where, l_ctr_counts)):
-            if counts % tile == 0: continue
-            n_alined = (counts+tile-1) & (0x100000-tile)
-            padding = n_alined - counts
-            l_ctr_counts[n] = n_alined
-
-            bas = mol._bas[m].copy()
-            bas[PTR_COEFF] = env_ptr
-            pad_bas.extend([bas] * padding)
-            pad_inv_idx.extend([n] * padding)
-
-            l = l_ctr[0]
-            nf = (l + 1) * (l + 2) // 2
-            coeff.extend([np.zeros((nf, nao_orig))] * padding)
-
-        inv_idx = np.hstack([inv_idx.ravel(), pad_inv_idx])
-        sorted_idx = np.argsort(inv_idx, kind='stable').astype(np.int32)
-        self.coeff = cp.asarray(np.vstack([coeff[i] for i in sorted_idx]))
-        assert self.coeff.shape[0] < 32768
-
-        max_nprims = uniq_l_ctr[:,1].max()
-        mol._env = np.append(mol._env, np.zeros(max_nprims))
-        if pad_bas:
-            mol._bas = np.vstack([mol._bas, pad_bas])[sorted_idx]
-        else:
-            mol._bas = mol._bas[sorted_idx]
-        assert mol._bas.dtype == np.int32
-
-        ## Limit the number of AOs in each group
-        if group_size is not None:
-            uniq_l_ctr, l_ctr_counts = _split_l_ctr_groups(
-                uniq_l_ctr, l_ctr_counts, group_size, tile)
+        mol, coeff, uniq_l_ctr, l_ctr_counts = group_basis(mol, self.tile, group_size)
+        self.mol = mol
+        self.coeff = cp.asarray(coeff)
         self.uniq_l_ctr = uniq_l_ctr
         self.l_ctr_offsets = np.append(0, np.cumsum(l_ctr_counts))
-
-        if mol.verbose >= logger.DEBUG1:
-            log.debug1('Number of shells for each [l, nprim] group')
-            for l_ctr, n, n8 in zip(uniq_l_ctr, l_ctr_counts_orig, l_ctr_counts):
-                log.debug1('    %s : %s -> %s', l_ctr, n, n8)
-
-        # PTR_BAS_COORD is required by nr_contract_jk.c
-        mol._bas[:,PTR_BAS_COORD] = mol._atm[mol._bas[:,ATOM_OF],PTR_COORD]
 
         # very high angular momentum basis are processed on CPU
         lmax = uniq_l_ctr[:,0].max()
@@ -515,112 +455,6 @@ class _VHFOpt:
         )
         return self
 
-def basis_seg_contraction(mol, allow_replica=1):
-    '''transform generally contracted basis to segment contracted basis
-    Kwargs:
-        allow_replica:
-            when angular momentum lower than (or equal to) this value, transform
-            the generally contracted basis to replicated segment-contracted basis.
-            By default, high angular momentum functions (d, f shells) are fully
-            uncontracted.
-    '''
-    bas_templates = {}
-    _bas = []
-    _env = mol._env.copy()
-    contr_coeff = []
-    aoslices = mol.aoslice_by_atom()
-    for ia, (ib0, ib1) in enumerate(aoslices[:,:2]):
-        key = tuple(mol._bas[ib0:ib1,PTR_COEFF])
-        if key in bas_templates:
-            bas_of_ia, coeff = bas_templates[key]
-            bas_of_ia = bas_of_ia.copy()
-            bas_of_ia[:,ATOM_OF] = ia
-        else:
-            # Generate the template for decontracted basis
-            coeff = []
-            bas_of_ia = []
-            for shell in mol._bas[ib0:ib1]:
-                l = shell[ANG_OF]
-                nf = (l + 1) * (l + 2) // 2
-                nctr = shell[NCTR_OF]
-                if nctr == 1:
-                    bas_of_ia.append(shell)
-                    coeff.append(np.eye(nf))
-                    continue
-                # Only basis with nctr > 1 needs to be decontracted
-                nprim = shell[NPRIM_OF]
-                pcoeff = shell[PTR_COEFF]
-                if l <= allow_replica:
-                    coeff.extend([np.eye(nf)] * nctr)
-                    bs = np.repeat(shell[np.newaxis], nctr, axis=0)
-                    bs[:,NCTR_OF] = 1
-                    bs[:,PTR_COEFF] = np.arange(pcoeff, pcoeff+nprim*nctr, nprim)
-                    bas_of_ia.append(bs)
-                else: # To avoid recomputation, decontract to primitive functions
-                    pexp = shell[PTR_EXP]
-                    exps = _env[pexp:pexp+nprim]
-                    norm = gto_norm(l, exps)
-                    # remove normalization from contraction coefficients
-                    c = _env[pcoeff:pcoeff+nprim*nctr].reshape(nctr,nprim)
-                    c = np.einsum('ip,p,ef->iepf', c, 1/norm, np.eye(nf))
-                    coeff.append(c.reshape(nf*nctr, nf*nprim).T)
-
-                    _env[pcoeff:pcoeff+nprim] = norm
-                    bs = np.repeat(shell[np.newaxis], nprim, axis=0)
-                    bs[:,NPRIM_OF] = 1
-                    bs[:,NCTR_OF] = 1
-                    bs[:,PTR_EXP] = np.arange(pexp, pexp+nprim)
-                    bs[:,PTR_COEFF] = np.arange(pcoeff, pcoeff+nprim)
-                    bas_of_ia.append(bs)
-
-            if len(bas_of_ia) > 0:
-                bas_of_ia = np.vstack(bas_of_ia)
-                bas_templates[key] = (bas_of_ia, coeff)
-            else:
-                continue
-
-        _bas.append(bas_of_ia)
-        contr_coeff.extend(coeff)
-
-    pmol = mol.copy()
-    pmol.cart = True
-    pmol._bas = np.asarray(np.vstack(_bas), dtype=np.int32)
-    pmol._env = _env
-    contr_coeff = scipy.linalg.block_diag(*contr_coeff)
-
-    if not mol.cart:
-        contr_coeff = contr_coeff.dot(mol.cart2sph_coeff())
-    return pmol, contr_coeff
-
-def _split_l_ctr_groups(uniq_l_ctr, l_ctr_counts, group_size=NAO_IN_GROUP,
-                        align=TILE):
-    '''Splits l_ctr patterns into small groups with group_size the maximum
-    number of AOs in each group
-    '''
-    l = uniq_l_ctr[:,0]
-    nf = l * (l + 1) // 2
-    _l_ctrs = []
-    _l_ctr_counts = []
-    for l_ctr, counts in zip(uniq_l_ctr, l_ctr_counts):
-        l = l_ctr[0]
-        nf = (l + 1) * (l + 2) // 2
-        max_shells = max(group_size//nf-align+1, align, 2)
-        max_shells = (max_shells + align - 1) & (0x100000-align)
-        if counts <= max_shells:
-            _l_ctrs.append(l_ctr)
-            _l_ctr_counts.append(counts)
-            continue
-
-        nsubs, remaining = counts.__divmod__(max_shells)
-        _l_ctrs.extend([l_ctr] * nsubs)
-        _l_ctr_counts.extend([max_shells] * nsubs)
-        if remaining > 0:
-            _l_ctrs.append(l_ctr)
-            _l_ctr_counts.append(remaining)
-    uniq_l_ctr = np.vstack(_l_ctrs)
-    l_ctr_counts = np.hstack(_l_ctr_counts)
-    return uniq_l_ctr, l_ctr_counts
-
 class RysIntEnvVars(ctypes.Structure):
     _fields_ = [
         ('natm', ctypes.c_uint16),
@@ -671,6 +505,7 @@ def init_constant(mol):
 def _make_tril_tile_mappings(l_ctr_bas_loc, tile_q_cond, cutoff, tile=TILE):
     n_groups = len(l_ctr_bas_loc) - 1
     ntiles = tile_q_cond.shape[0]
+    tile_mask = cp.tril(tile_q_cond > cutoff)
     tile_mappings = {}
     for i in range(n_groups):
         for j in range(i+1):
@@ -681,9 +516,7 @@ def _make_tril_tile_mappings(l_ctr_bas_loc, tile_q_cond, cutoff, tile=TILE):
             j0 = jsh0 // tile
             j1 = jsh1 // tile
             sub_tile_q = tile_q_cond[i0:i1,j0:j1]
-            mask = sub_tile_q > cutoff
-            if i == j:
-                mask = cp.tril(mask)
+            mask = tile_mask[i0:i1,j0:j1]
             t_ij = (cp.arange(i0, i1, dtype=np.int32)[:,None] * ntiles +
                     cp.arange(j0, j1, dtype=np.int32))
             idx = cp.argsort(sub_tile_q[mask])[::-1]
